refactor(output): extract read_body_capped helper to http_util module
Both `output http` and `output otlp_http` need to bound how many
bytes of an error response body they read into memory. Today only
http.rs has the cap; otlp_http still calls `resp.text().await` and
buffers the entire body before trimming, which a multi-MB error
response can OOM-vector. Promoting the helper to a shared module
unblocks reusing the same invariant in otlp_http (next commit) and
sets the place where any third HTTP sink should pick it up too.

The 4 KiB cap, the use of `Response::chunk()` (no `stream` reqwest
feature dependency), and the trim semantics are unchanged. The
misleading comment on the mid-chunk break path is corrected: it
previously claimed the peer's connection would be returned to the
pool, but reqwest/hyper actually closes the underlying TCP
connection when the `Response` is dropped without reaching EOF.
The new wording acknowledges that and notes why the trade-off is
acceptable (the error path is rare, bounded memory matters more
than reusing a connection to a peer that just failed).

Pure refactor + comment fix; no behaviour change in `output http`.
